+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **New Features**
 * replaces non-standard-conformant idTypes with default (IRI)
 * removes View elements if present
+* now correctly converts MultiLanguageProperty value
 
 **Internal changes & Bugfixes**
 * Fixed handling idType ID_SHORT